refactor: add appropriate namespaces for classes
