+++ index.php
@@ -2,8 +2,8 @@
 
 use Xvlvv\Enums\Action;
 use Xvlvv\Enums\Status;
-use Xvlvv\Task;
-use Xvlvv\TaskStateManager;
+use Xvlvv\Models\Task;
+use Xvlvv\Services\TaskStateManager;
 
 require_once __DIR__ . '/vendor/autoload.php';
